+++ src/main/kotlin/platform/AbstractModule.kt
@@ -24,12 +24,17 @@ import com.demonwav.mcdev.facet.MinecraftFacet
 import com.demonwav.mcdev.insight.generation.EventListenerGenerationSupport
 import com.demonwav.mcdev.inspection.IsCancelled
 import com.intellij.openapi.module.Module
+import com.intellij.openapi.util.Key
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiExpression
 import com.intellij.psi.PsiMethod
 import com.intellij.psi.PsiMethodCallExpression
 import com.intellij.psi.PsiReferenceExpression
+import com.intellij.psi.util.CachedValue
+import com.intellij.psi.util.CachedValueProvider
+import com.intellij.psi.util.CachedValuesManager
+import com.intellij.psi.util.PsiModificationTracker
 import javax.swing.Icon
 
 abstract class AbstractModule(protected val facet: MinecraftFacet) {
@@ -43,6 +48,21 @@ abstract class AbstractModule(protected val facet: MinecraftFacet) {
 
     open val icon: Icon?
         get() = moduleType.icon
+
+    protected open fun computeModIds(): List<String> = emptyList()
+
+    private val modIdsCacheKey = Key.create<CachedValue<List<String>>>("modIds for ${javaClass.simpleName}")
+
+    /**
+     * Returns a list of mod IDs for this module. Must be called in a read action!
+     */
+    val modIds: List<String>
+        get() {
+            val provider = CachedValueProvider {
+                CachedValueProvider.Result(computeModIds(), PsiModificationTracker.MODIFICATION_COUNT)
+            }
+            return CachedValuesManager.getManager(project).getCachedValue(module, modIdsCacheKey, provider, false)
+        }
 
     open val eventListenerGenSupport: EventListenerGenerationSupport? = null