+++ src/main/kotlin/platform/neoforge/util/NeoForgeConstants.kt
@@ -31,4 +31,5 @@ object NeoForgeConstants {
     const val MCMOD_INFO = "mcmod.info"
     const val PACK_MCMETA = "pack.mcmeta"
     const val MODS_TOML = "neoforge.mods.toml"
+    const val MODS_TOML_PATH = "META-INF/$MODS_TOML"
 }